Refactor page classes by moving them into new classes. Update imports
