+++ src/main/java/com/unityTest/courseManagement/models/api/response/page/CourseAttributePage.java
@@ -1,4 +1,4 @@
-package com.unityTest.courseManagement.models;
+package com.unityTest.courseManagement.models.api.response.page;
 
 import com.unityTest.courseManagement.entity.CourseAttribute;
 import io.swagger.annotations.ApiModel;
